LC-equal-sum-grid-partition-i [Runtime: N/A; Memory: N/A]

+++ matrix/equal-sum-grid-partition-i.cpp
@@ -24,7 +24,7 @@ public:
         }
 
 
-        int leftSum = 0;
+        long long leftSum = 0;
         for(int j =0; j<n-1; j++){
             for(int i=0; i<m; i++){
                 leftSum += grid[i][j];
